backend/requirements.txt
fastapi
uvicorn
python-multipart
youtube-transcript-api
google-generativeai
sentence-transformers
chromadb
python-dotenv
pydantic
requests
numpy
langchain
langchain-community
langchain-google-genai
pytube
regex
tiktoken